refactor(web): add high level event emitter
